+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Preserve active sorting and filtering when selecting publication in publications list.
 - Preserve active sorting and filtering when viewing images in a facsimile collection and returning to facsimile collection list.
 - Custom table component: move data subscription logic to ngOnInit.
+- Custom table component: move static filterable columns initialization outside data subscription logic.
 - Deps (dev): update `angular-eslint` to 18.4.2.
 
 ### Fixed